Add "bank helper" userscript
`bank_helper.js` implements a single feature, at the "Bank" area only:

Whenever your credits "on hand" are greater than the amount you'd always
like to keep at hand, it'll prefill the "deposit" textbox with the
excess cash.

Note that this does *not* automatically deposit the credits.
It simply fills the box for you.