Added "Heavy Dö-maru" to special-case slugs list.
Technically, this is also a typo; it should be "Heavy Dō-maru". However, if staff were to change the name without also updating the slug (sometimes they do, sometimes they don't), it would disagree with the existing "ō" -> "o" rule -- hence adding this as a special case instead of a new "ö" rule.