Update discreet_keyboard_navigation for 2021-04-23 update
Updated to take an approach similar to discreet_helper which might be made
redundant (sorry). Uses MutationObserver to watch for added child elements
and adds links which are then used by the keyboard shortcut. I use the same
data label as discreet_helper so that if both are enabled they don't conflict.

This script does not (yet?) modify the discreet work link upon completion
of the discreet work.